+++ R/breaks.R
@@ -419,6 +419,7 @@ brk_manual <- function (breaks, left_vec) {
   }
 }
 
+
 #' @rdname chop_fn
 #' @export
 #' @order 2
@@ -438,6 +439,58 @@ brk_fn <- function (fn, ...) {
     breaks <- create_extended_breaks(breaks, x, extend, left, close_end)
 
     breaks
+  }
+}
+
+
+#' @rdname chop_spikes
+#' @export
+#' @order 2
+brk_spikes <- function (breaks, n = NULL, prop = NULL) {
+  assert_that(
+    is.number(n) || is.number(prop),
+    is.null(n) || is.null(prop),
+    msg = "exactly one of `n` and `prop` must be specified as a scalar numeric"
+  )
+  if (! is.function(breaks)) breaks <- brk_default(breaks)
+
+  function (x, extend, left, close_end) {
+    n <- n %||% (length(x) * prop)
+    unique_x <- unique(x)
+    x_counts <- tabulate(match(x, unique_x))
+    spikes <- unique_x[x_counts >= n]
+
+    breaks <- breaks(x, extend, left, close_end)
+    break_elements <- unclass_breaks(breaks)
+    left_vec <- attr(breaks, "left")
+
+    # We sort spikes in decreasing order so that when we add elements,
+    # earlier elements remain in place.
+    spikes <- sort(spikes, decreasing = TRUE)
+    for (spike in spikes) {
+      # We could use match() here to go faster, or even put it outside the loop.
+      match_location <- which(spike == break_elements)
+      n_matches <- length(match_location)
+      # If two break elements match the spike, it's already a singleton:
+      # we don't need to do anything.
+      if (n_matches >= 2L) next
+      if (n_matches == 1L) {
+        # We turn the single matching break into a singleton and make sure
+        # that left is c(TRUE, FALSE)
+        break_elements <- append(break_elements, spike, after = match_location)
+        left_vec <- append(left_vec, FALSE, after = match_location)
+        left_vec[match_location] <- TRUE
+      } else {
+        # We add a singleton break at `spike`
+        insert_location <- quiet_max(which(spike > break_elements))
+        if (insert_location <= 0) insert_location <- 0
+        break_elements <- append(break_elements, rep(spike, 2),
+                                 after = insert_location)
+        left_vec <- append(left_vec, c(TRUE, FALSE), after = insert_location)
+      }
+    }
+
+    create_breaks(break_elements, left = left_vec)
   }
 }