+++ NEWS.md
@@ -1,9 +1,11 @@
 # santoku (development version)
 
+* Experimental: the new `chop_spikes()` function gives common values of `x` their own
+  singleton break.
 * Singleton breaks are not labelled specially by default in 
   `chop_quantiles(..., raw = FALSE)`. This means that e.g. if the 10th and 20th
   percentiles are both the same number, the label will still be `[10%, 20%]`.
-* Document `single = NULL` explicitly in `lbl_*` functions.
+* `single = NULL` has been documented explicitly in `lbl_*` functions.
 
 # santoku 1.0.0